Change parameters and coordinates of simulation to float
Refactor paramters in two different vectors, declared as global
variables. Parameters P, V, W and L are kept as integers, while
parameters Np, Nv, Dp, Dv, Vp, Vv and t are now an array of floats.

Moreover, the functions advance_person and advance_vehicle now use a
more comprehensible enum to express the direction.

As TODOs, we need to implement an euality check for floats in the
print_region function (for now, it should not affect the correctness
of the simulation); and the coordinates toroidal wrap of functions
advance_person and advance_vehicle only works if the wrap is at most
one (see comments at beginning of functions)