Add noise calculations for grid of virtual sensors (TESTING NEEDED)
- The simulation now calculates the noise levels produced by people
and vehicles, using the functions intensity, intensity_at_distance and
decibels

- To ease the computation, the dB intenisty is converted in W
intensity (so that it can be summed). At the end of the computation,
it is conVerted back in dB

- The function distance calculates the euclidean distance between an
agent and a sensor

- Function do_simulation has been enriched to produce the noise data,
using function produce_sensor_data (see next point)

- function produce_sensor_data iterates through each sensor point,
accumulationg the noise produced by nearby people and vehicles. Then,
the noise level in dB is computed

- TODO: send data to coordinato process and the Spark cluster
- TODO: test the newly introduced functions