[BUGFIX] Fix sound intesity going to infinity + debug helper string
Previously, the sound intensity was calculated as (intensity /
distance^2). This produced infinity values when distance was 0. Now
the calculation is (intensity / (distance + 1)^2), effectively placing
the nominal noise level at source position.

Moreover, an example command to launch a typical session of gdb
correctly in Emacs is provided in file gdb_run_string.txt