Add TCP network connection
Now the simulator accepts as command line arguments the address and
the port on which the Kafka producer is listening, creates a TCP
connection to it, and sends the data strings to it. The helper scripts
have been modified for example. The program works single-threaded and
in localhost, but IT STILL MUST BE TESTED ON THE REAL SETTING