[Akka] Add first fault tolerance mechanism
The MainDispatcher now sends new tasks to JobSupervisorActors, that
monitor executions of JobWorkerActor. If a failure occurs (for now
modeled with a pseudorandom number), then the JobWorkerActor stash the
message, restarts, and unstash the message. However, no synchronous
request is implemented, and the system may potentially shutdown before
each job has terminated