[Akka] Change to SmallestMailboxPool usage (load balancing)
Now the cluster uses a SmallestMailboxPool to instantiate the other
actors. The couple supervisor/worker is still in use, to actually
implement fault tolerance. The system now automatically tries to route
new messages to the actor which is free and/or with the smallest
queue. NOTE: synchronous replies are still not supported, the project
works if the actor system is not terminated before the end of the
pending computations