[Akka] Implement JSON reading of job duration
Now the main method of class ClusterStarter reads the file JSON
located in src/main/resources/job_list.json, extracts its data in a
vector, and subsequently in an HashMap. The messages that encapsulate
a job request now contain also the duration of the job, taken from
that map. The JobWorkerActors use now that value (in ms) as wait time.