[Kafka] Add basic "Logging and analysis" point 1
Now ClusterStarter starts a new Thread that performs the point 1 of
the required analysis. It reads the topic completedJobs without
committing its offsets, performs the analysis, then unsubscribes from
the topic and repeats. This way it always reads from the start. May be
imporoved using the seek() methods.
In addition, the topic retention times have been expanded to one
month, and each topic is created with 5 partitions