[Kafka] Add basic "Logging and analysis" point 2
Now the cluster keeps track of the number of pending jobs, computed as
the difference between the number of records on the pendingJobs and
the completedJobs topics. The analysis may be slightly off, when jobs
are submitted/completed before or after the analysis thread has run.
Due to this error, if the number of pending jobs gets computed as < 0,
it is then resetted to 0.
NOTE: if a malformed message is published on a topic, it is still
counted for the total! This may lead to zombies pending jobs that get
never deleted