[Akka] Add logging (in messages) of number of starts for each job
Now when a job is started, the nOfFailures is piggybacked in the
JobTaskMessage, and used each time the message it is stashed/unstashed
(i.e., at failures). At each failure, nOfFailures is incremented with
incFailures(). Upon completing, the nOfStarts is computed as
nOfFailures + 1, and inserted in the JobCompletedMessage, then echoed
on System.out