[Akka][Fixup] Correctly indicate the sender of the JobTaskMessage
In JobSupervisorActor::idleUntilNewJob(), the JobTaskMessage sent to
the JobWorkerActor was erroneously created with ActorRef.noSender().
Now it is correctly constructed with the JobSupervisorActor as the
correct sender.
NOTE: some debug prints to System.err have been added
NOTE: the system is still slow to pick up uread Kafka events,
accumulated while the actors were offline