[Akka] Complete basic version of Logging and analysis part 3
The CompletedAnalysisThread, after completing its analysis, triggers
RetriesAnalysisActor with a StartAnalysisMessage (in fact, both of
them should run once in a minute).
RetriesAnalysisActor, when it receives a JobStatisticsMessage, it
enqueues it in a deque. When it performs the analysis, it first
discards all messages older than a minute; then, it computes the
average number of starts, with the help of a Map of a private
class (JobCounts). After the analysis, the map entries are resetted to
0.
NOTE: the debug prints are still present
NOTE: since the analysis runs once in a minute, it could be possible
to simply clear all the deque after finishing the analysis
NOTE: the analysis doesn't read from Kafka: therefore, it doesn't
accounts for completions published technically within the latest
minute, but before the cluster started running. However, the resulting
discrepancy is limited to one minute